requirements.txt
langchain==0.3.12
langchain-google-genai==2.0.8
langchain-community==0.3.12
langchain-chroma==0.1.4
neo4j==5.27.0
chromadb==0.5.23
streamlit==1.41.1
streamlit-agraph==0.0.45
pydantic==2.10.4
python-dotenv==1.0.1
tenacity==9.0.0
google-generativeai==0.8.3
plotly
